+++ src/api/providers_custom.js
@@ -1,9 +1,24 @@
-import { getChatName, getCustomAPIInfo } from "#root/src/components/preferences/manageCustomAPIs.jsx";
-import { CustomOpenAIProvider } from "#root/src/api/Providers/special/custom_openai.jsx";
 import { removePrefix } from "#root/src/helpers/helper.js";
+import { Storage } from "#root/src/api/storage.js";
 
 import { providers_info } from "#root/src/api/providers_info.js";
 import { chat_providers_names } from "#root/src/api/providers_react.jsx";
+
+import { CustomOpenAIProvider } from "#root/src/api/Providers/special/custom_openai.jsx";
+
+const customAPIsStorageKey = "customAPIs";
+export const getCustomAPIInfo = async (url) => {
+  const data = JSON.parse(await Storage.read(customAPIsStorageKey, "{}")) || {};
+  if (url) return data[url];
+  return data;
+};
+export const updateCustomAPIInfo = async (data) => {
+  await Storage.write(customAPIsStorageKey, JSON.stringify(data));
+};
+
+export const getChatName = (api) => {
+  return api.name || api.url;
+};
 
 export const customAPItoKey = (api, model) => {
   // Notice that we only need to use the url and model to uniquely identify a custom API entry.
@@ -25,7 +40,7 @@ export const initCustomAPIs = async () => {
     const models = api?.models ?? [];
     for (let model of models) {
       const key = customAPItoKey(api, model);
-      providers_info[key] = { provider: CustomOpenAIProvider, url: url, model: model };
+      providers_info[key] = { provider: CustomOpenAIProvider, url: url, model: model, stream: true };
       chat_providers_names[key] = `${getChatName(api)} (${model})`;
     }
   }
